Reorganize repository and update README to reflect changes.
This refactoring moved files around, though I haven't gone through all the noteboooks and scripts to update the references from the files, so that will be a next step. I tried to place config files and scripts in their respective correct directories and get the organization of my repository more sensible.