Make changes to planning.md file to reflect current progress
I initially made my backlog, ice-box, etc... in the README but have moved it due to clutter. I just went through and updated my progress on various tasks, changed the columns in the tables a bit, and added new tasks which have come up as well as reorganized the theme/categories a bit.