get LDA model script
I made some changes to my LDA model notebook, converted it to a script, and decided I would create this model after running some of the other scripts which filter and stem my plot summaries. Now this script will create a mkl object which will be used eventually to create features for plot summaries which are as many as LDA topics I have with my model.