Work on web server application and returning results
This commit made major changes to the app.py file which runns my flask app. It has 2 routes, a home page
and a results page where results are displayed.

I created a main .html template and inherited the structure to my others.

I allowed for a form to be on the home page where a new plot summary can be entered and results will be gotten from that.